requirements.txt
astroid==2.3.3
azure-functions==1.2.0
colorama==0.4.3
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
openpyxl==3.0.3
pandas==1.0.3
pandas-schema==0.3.5
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
